AnnotationFlow extracted from AnnotationJob
- This allows tests for any job dependant on AnnotationFlow to autowire JobLauncherTestUtils
- JobLauncherTestUtils autowired into AggregatedVcfJobTest
