+++ src/test/java/uk/ac/ebi/eva/pipeline/jobs/AggregatedVcfJobTest.java
@@ -25,11 +25,8 @@ import org.opencb.opencga.storage.core.variant.VariantStorageManager;
 import org.opencb.opencga.storage.core.variant.adaptors.VariantDBAdaptor;
 import org.opencb.opencga.storage.core.variant.adaptors.VariantDBIterator;
 import org.springframework.batch.core.*;
-import org.springframework.batch.core.launch.JobLauncher;
-import org.springframework.batch.core.repository.JobRepository;
 import org.springframework.batch.test.JobLauncherTestUtils;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.beans.factory.annotation.Qualifier;
 import org.springframework.boot.test.IntegrationTest;
 import org.springframework.test.context.ContextConfiguration;
 import org.springframework.test.context.junit4.SpringJUnit4ClassRunner;
@@ -56,20 +53,15 @@ import static uk.ac.ebi.eva.test.utils.JobTestUtils.getTransformedOutputPath;
  */
 @IntegrationTest
 @RunWith(SpringJUnit4ClassRunner.class)
-@ContextConfiguration(classes = {JobOptions.class, AggregatedVcfJob.class, VariantAggregatedConfig.class})
+@ContextConfiguration(classes = {JobOptions.class, AggregatedVcfJob.class, VariantAggregatedConfig.class, JobLauncherTestUtils.class})
 public class AggregatedVcfJobTest {
 
     @Autowired
-    @Qualifier("aggregatedJob")
-    public Job job;
+    private JobLauncherTestUtils jobLauncherTestUtils;
+    
     @Autowired
     private JobOptions jobOptions;
-    @Autowired
-    private JobLauncher jobLauncher;
-    @Autowired
-    private JobRepository jobRepository;
 
-    private JobLauncherTestUtils jobLauncherTestUtils;
     private String input;
     private String outputDir;
     private String compressExtension;
@@ -161,11 +153,6 @@ public class AggregatedVcfJobTest {
     @Before
     public void setUp() throws Exception {
         jobOptions.loadArgs();
-
-        jobLauncherTestUtils = new JobLauncherTestUtils();
-        jobLauncherTestUtils.setJob(job);
-        jobLauncherTestUtils.setJobLauncher(jobLauncher);
-        jobLauncherTestUtils.setJobRepository(jobRepository);
 
         input = jobOptions.getPipelineOptions().getString("input.vcf");
         outputDir = jobOptions.getPipelineOptions().getString("output.dir");
